fix Twisted-10.2 compatiblity: stop using strports.service() . Fixes #167.
Replaced it with a manual parser that should handle the subset of the
strports format we know how to deal with (TCP on IPv4, with an optional
interface= specification). In Twisted-10.2 (just released an hour ago),
strports.service() returns something new (not the usual TCPServer), that does
not have a _port attribute.