run-command client: workaround a Twisted StandardIO half-closed-stdio bug
When our parent closes out stdout, then writes to our stdin, a race occurs.
If the stdout-is-closed appears first, Twisted will close the whole
connection, and we'll never see the stdin.

The workaround tweaks some internals to make this work better. I'm not sure
how robust this will be: fixing it upstream would be better, of course.