track TCP connections going away separately from Brokers becoming unused
Also don't assume loseConnection() is synchronous, specifically when dropping
a connection to replace it with a new one. In Twisted-11.1 (with the new TLS
changes), connectionLost() happens some time *after* loseConnection() is
called. We still do brokerDetached() in shutdown(), but Tub.stopService() is
stalled until after all the connections have finally fired connectionLost().