setup.py: remove extras["secure_connections"], as it isn't reliable
Since extras_require is only usable under setuptools, and PyOpenSSL is a
soft dependency, and setuptools is explicitly not a dependency, other
Python packages cannot reliably require that we have PyOpenSSL without
requiring it themselves. So this extras_require is not reliable enough
to retain. Thanks to Zooko for the patch to remove it.

Also added a README.packagers to explain to both OS-level package
authors and python-package authors what they need to depend upon to get
secure connections from Foolscap.

Closes #174