setup.py: require setuptools
I've avoided this for many years, because setuptools has historically
caused a lot of pain and suffering, but I think modern setuptools is a
lot better. And being able to say "pip install foolscap" is great.

This form still uses scripts= on non-windows. And this change removed
the tests_require= setting.. I think that was redundant.