Use client Endpoints for outbound connections. Needs Twisted>=10.1
This also simplifies the TubConnector a bit, now that we can use
cancellable Deferreds to abandon the losing connections.

setup.py's install_requires= was bumped to require Twisted-10.1.0, since
that's where TCP4ClientEndpoint first got added. If we could use
HostnameEndpoint (added in 13.2.0) we'd probably get IPv6 support, but
we need to wait for twisted#8014 to get fixed first (to avoid annoying
error messages when using multiple connection hints).

refs ticket:236