remove shared Listeners, Tub.listenOn(listener), .clone()
The thorny who-owns-the-Listener problem described in ticket:239 was
easiest to solve by removing the Foolscap feature of allowing Listeners
to be shared between multiple Tubs. With this change, each Listener has
exactly 0 or 1 parent Tub (zero-Tub Listeners aren't useful now, but
might be useful in the future when we implement redirects or relays).

It was also necessary to remove Tub.clone(), since it necessarily
produced shared Listeners.

We also remove Listener.addTub and .removeTub, since they're no longer
used, and simplify Listener internals (Listeners are now Services, and
the listening socket is a service child). This patch also removes the
static global "Listeners" array, which was only used for testing, but
felt like a bad idea anyways. And to finish the exercise, we've excised
the tests which exercised the excised feature.

refs ticket:239