test unreachable (location-less) Tubs
Make sure that a client (without a location) can send one of its own
objects to a server (with a location), and that the server can then call
methods on that object.

We also confirm that the tub.setOption("logport-furlfile") doesn't
explode when you don't set a location: the logport.furl file just never
gets written.

Note that the way you tell Foolscap that you want an unreachable
client-only Tub is to *not* call tub.listenOn() or tub.setLocation().

Closes ticket:208.