unreachable gifts should trigger an error upon receipt
Arrange for a gift to arrive at the recipient as an
unresolveable (their-reference) sequence, causing the error to occur on
the recipient. I tried raising a Violation during slicing
instead (making it error occur on the sender), but that got the slicing
engine confused and the connection was aborted because it fell out of
sync.

refs ticket:208