switch to HostnameEndpoint, enable DNS-based IPv6 connections
This doesn't enable IPv6 completely, but it should allow us to connect
to DNS-based hints which happen to resolve to IPv6 addresses (previously
we could only handle hostnames that use IPv4).

It does not yet:

* handle hints like "tcp:[12AB:IPV6:ADDR::STUFF]:PORT"
* make it possible to listen on v6 ports
* auto-detect local v6 addresses

This bumps our Twisted dependency to >=16.0.0, since we need the fix to
Twisted bug 8014, to avoid Unhandled Error In Deferred when a FURL
contains multiple hints and the losers are cancelled. This also gets us
Twisted >= 13.2.0, which is where HostnameEndpoint was introduced.

refs ticket:155 (but does not close it, as those remaining three issues
still need to be resolved)