tor: don't pass reactor into handler constructor
Instead, get it from hint_to_endpoint().

My original thought was that the handler constructor function (e.g.
tor.launch()) should return a Deferred that fires with the handler,
which would give it a chance to start the Tor daemon right away (instead
of waiting for the first hint to be processed). To enable that, the
constructor would need access to a reactor.

However then Tahoe (and other applications) could not set up their Tubs
synchronously, since they'd need to wait for the Deferred to fire before
they could add the handler.

I don't think we lose much by waiting for the first hint to do one-time
setup.