tor: tolerate non-numeric config.SocksPort values
When doing manual testing with a real Tor server, I got back a
config.SocksPort with two lines, one of which was
"unix:/var/run/tor/socks WorldWritable", which confused the parser. This
patch improves the parser to ignore non-numeric lines. In the future, we
could probably build an endpoint from the unix line and use that
instead.