tor: clean up implementation, update tests
The internal _connect() method now yields the SOCKS endpoint which all
Tor handlers use, instead of storing it as self._socks_endpoint. When
useful, we stash the descriptor from which we create that endpoint, so
tests can look at it later.

Tests that can call the real txtorcon methods do so: others mock out the
necessary parts of txtorcon so we avoid actually talking to Tor (or
trying to launch it).