allocate_tcp_port: avoid loopback-bound sockets
I thought I'd fixed this, but I'm still seeing occasional failures,
where another process has bound a socket to 127.0.0.1, and
allocate_tcp_port() gives it to me anyways. I wouldn't be surprised if
the code that's using the socket is binding it to lo0 too, otherwise it
might not cause a problem.

I think the fix is for allocate_tcp_port() to attempt to bind a
socket (with the proposed port number) to both 0.0.0.0 and then
127.0.0.1, seeing if either gives us an error.

(re) closes ticket:258