IPv6: accept IPv4-mapped and local-scope/zone-id too
These are kind of fringe use cases:

* if you've got an IPv4-mapped address, you might as well just use
  "tcp:1.2.3.4:9900"
* site-local addresses are deprecated
* link-local addresses aren't useful to anyone outside your subnet

but they're legitimate IPv6 addresses, and it's probably easier to
accept them here, than to ask applications to filter them out or convert
them.

This patch also adds some comments to connections/tcp.py, now that the
regexps are a bit more complex.

closes ticket:155