getConnectionInfoForFURL: tolerate unparseable FURLs
There's a Tahoe test case that launches a node with a FURL of "None" (a
string), and was depending upon being able to start a Reconnector against
this without complaints. Reconnector._connect() now parses the FURL (as a
side-effect of stashing the initial ConnectionInfo object), which threw a
ValueError when "None" didn't start with "pb://".

This changes getConnectionInfoForFURL() to return None for unparseable FURLs,
which makes sense anyways, and happens to keep that test case working.