connector: report unrecognized failures better
Previously this put str(exception) into the connector status, but when that
is empty, we now fall back to the repr. For many failed-to-connect
exceptions (e.g. txsocksx errors), the classname is the only useful piece of
information we've got, and that's only included in the repr, not the str().