fix unreliable test
CrossfireReverse.test4 was failing about 50% of the time, seemingly more
frequently now after an unrelated change. This test needed rewriting badly:
it depended upon a couple of fixed-duration stalls and a debug hook in
Negotiation that paused one of the messages for a similar duration. The new
test completely blocks the message and has no time-based stalls.

More work needs to be done (all the Crossfire tests should be switched to
this non-time-based approach), but it's a large rewrite and I haven't
finished it yet.