fix Gatherer: wire-side Events need bytes keys, gathered incidents need magic
The foolscap.logging.interfaces schema requires both Events and Headers (as
used by trigger= arguments) to be dictionarys with keys that are bytes ("str"
in py2). But since we're pulling these from a JSON object, we get unicode.
The fix is to add a function that encodes everything to ASCII before sending
over the wire. This is safe because all these keys are static strings (added
by foolscap's logging functions). The application-provided data, which could
be of any type, lives in the values, not the keys.

Also, the incident files written by the gatherer were lacking the version
prefix (aka MAGIC), so the fix was to write this just after opening the file
for writing, before the header dict is written.