logs: fix serialization of unserializable objects
If an error occurs during serialization (e.g. one of the event arguments
cannot be expressed in JSON), we now use a placeholder object that holds its
repr(). If repr() fails, we use a second type placeholder that holds the
repr() of the exception. If the repr() of the exception fails, we use a third
type of placeholder that merely contains a message of sympathy and hope for
the unlucky developer.