test_pb: local logged failures are no longer CopiedFailure objects
We used to replace Failures with CopiedFailures in the local circular
buffers, because pickle would put the original Failure's exception type into
the serialized objects (in an Incident file), which would require the
subsequent loading application (e.g. flogtool dump) to try to import the same
exception, which would probably fail.

Now that Failures are being transformed by the JSON serializer, we don't need
to muck with the circular buffers: Failures remain as-is until they get
serialized.