test_reconnector: improve reliability
This test depended upon the stall() values being long enough to allow a
connection attempt (and loss, and reattempt) to complete first. On slower
systems, this wasn't enough.

I've added poll() calls to make sure the cycle is complete before allowing
the test to proceed. To improve test coverage, I've left the poll()s in
place (so most of the time we'll exercise the timer-expiration and
reconnection code too), but they should no longer be critical to allowing the
test to pass.

A future project is to replace all of the polls() and stalls() with explicit
timer manipulation, ala Twisted's "clock" pattern (e.g. the Reconnector
doesn't get a real reactor, it gets a controlled Clock object, and the test
deliberately increments it at a known pace to examine the intermediate
state).