delete UnsafeBanana
I built this to offer Banana-style serialization of everything that Pickle
could handle: modules, classes, instances, methods. I never used it, and it's
just as horribly unsafe as pickle, so I didn't recommend anyone else to use
it either. The tricks it uses to create new instances aren't trivial to port
to Python 3, so it seems best to just delete it entirely.