fix test of test_negotiate by fixing connection.py
TubConnector.checkForFailure is called after each event which might indicate
that the overall connection attempt to a given Tub has failed. It looks to
see if any more work might be done: if so, we haven't failed yet. But if all
our hopes are dashed, it declares failure.

It had an escape clause, however: if one of the outbound connection attempts
was rejected by the other side as a "Duplicate connection", that suggests
that the other side thinks it has a valid connection to us. In that case,
checkForFailure refrained from declaring failure just yet, in the hopes that
we'll hear about this duplicate connection shortly. I think I added this to
avoid a spurious failure in the face of two sides simultaneously connecting
to each other.

However the escape clause was broken: it used a function named isSubstring()
with the arguments backwards, so it could never ever return True. When I
fixed that use of isSubstring, a number of unit tests started failing. I
don't have the patience to figure out the details now, so I just disabled the
clause.