fix rest of test_logging, adds incompatibility for Gatherer
The "logport" is a Foolscap object that enables one process to publish
structured log events and "Incidents" (causally/temporally related
collections of events) to another process. Foolscap comes with tools to use
this: `flogtool tail` will connect to a logport and show all events,
`flogtool gatherer` creates a process that reads and stores all events, and
more.

The protocol spoken over the logport includes various string-bearing
structures, including Incident names, and the fixed keys of various
dictionaries, in addition to strings usually being present in the
application-provided log events themselves.

The previous code enforced a "schema" on the logport messages, requiring the
strings to be py2-style bytestrings. The new code removes that constraint,
allowing them to be anything. In practice, they will probably be
python-native strings (so bytes on py2, and unicode on py3).

Thus if a py2 process exposes a logport, and the follower/gatherer is running
py3, at best the results will be messy (I'd expect a lot spurious `b"foo"` in
the output, rather than normal `foo` strings), and at worst it is likely that
connections or messages will be dropped for violating the old constraints.

I hope to improve the behavior of the new code in the future, to avoid the
messy bytes representations. But it will take a lot more work to enable
interop between py2 and py3 processes (not impossible, the number of strings
sent *into* the logport is small, and the new py3 side might conceivably
observe the other side to tell what sorts of strings it will accept).

This enables all of test_logging to pass, although I must still manually
exercise the various CLI tools to see if they actually work.