logging: app versions are bytes on wire, improve format_message
Log events on the wire are dictionaries, with bytes as keys. This adds code
to convert those to native string keys upon receipt, so basic stuff like 'if
"time" in e' still works. We leave the values of those dictionaries alone:
applications can add arbitrary arguments (but the *names* of those arguments,
which are dictionary keys too, are converted).

`format_message()`, which makes a best-effort attempt to convert an event
dictionary into a printable string, was updated to tolerate more situations,
and to always return text rather than a native string. All places that use
format_message now expect text, even on py2.